Fix response time test

+++ backend/__tests__/non_functional_requirements/performance.test.ts
@@ -1,13 +1,12 @@
 describe("Performance test", () => {
-  beforeAll(async () => {
-    // sleep 2 seconds to allow proper setup
-    await new Promise((resolve) => setTimeout(resolve, 2000));
-  });
+  const USER_ID = "abcdefghijklmnopqrstuvwxyz";
+  const MAX_RESPONSE_TIME = 2000; // 2 seconds
 
   // Justification: A user may decide to generate a route and recipes, but then decide to delete both
   // route and recipes if they find them unsatisfactory, which should possible to do quickly
   test("Single route, 3 stops", async () => {
-    const start = Date.now();
+    // check route generate response time
+    var start = Date.now();
     const route_response = await fetch(
       `${process.env.GATEWAY_BASE_URL}/routes`,
       {
@@ -16,18 +15,25 @@ describe("Performance test", () => {
           "Content-Type": "application/json",
         },
         body: JSON.stringify({
-          userID: "test_person3",
+          userID: USER_ID,
           origin: "Vancouver",
           destination: "Toronto",
-          numStops: 1,
+          numStops: 3,
         }),
       }
     );
+    var end = Date.now();
 
     if (route_response.ok) {
+      var duration = end - start;
+      console.debug(`Route Execution time: ${duration}ms`);
+      expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
+
       const data = await route_response.json();
       const tripID = data.tripID;
 
+      // check recipe generate response time
+      start = Date.now();
       const recipe_response = await fetch(
         `${process.env.GATEWAY_BASE_URL}/recipes`,
         {
@@ -37,13 +43,20 @@ describe("Performance test", () => {
           },
           body: JSON.stringify({
             tripID,
+            userID: USER_ID,
           }),
         }
       );
+      end = Date.now();
+      duration = end - start;
+      console.debug(`Recipe Execution time: ${duration}ms`);
+      expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
 
       const recipe_data = await recipe_response.json();
       expect(recipe_data).not.toBeNull();
 
+      // check route delete response time
+      start = Date.now();
       const route_teardown = await fetch(
         `${process.env.GATEWAY_BASE_URL}/routes/${tripID}`,
         {
@@ -53,10 +66,18 @@ describe("Performance test", () => {
           },
         }
       );
+      end = Date.now();
+
       if (!route_teardown.ok) {
         throw new Error(`HTTP error! Status: ${route_teardown.status}`);
       }
 
+      duration = end - start;
+      console.debug(`Route Teardown Execution time: ${duration}ms`);
+      expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
+
+      // check recipe delete response time
+      start = Date.now();
       const recipe_teardown = await fetch(
         `${process.env.GATEWAY_BASE_URL}/recipes/${tripID}`,
         {
@@ -66,23 +87,24 @@ describe("Performance test", () => {
           },
         }
       );
+      end = Date.now();
       if (!recipe_teardown.ok) {
         throw new Error(`HTTP error! Status: ${recipe_teardown.status}`);
       }
+
+      duration = end - start;
+      console.debug(`Recipe Teardown Execution time: ${duration}ms`);
+      expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
     } else {
       console.error("Request failed with status:", route_response.status);
     }
-
-    const end = Date.now();
-    const duration = end - start; //begin timing test assuming that operation succeeded
-    console.debug(`Route Execution time: ${duration}ms`);
-    expect(duration).toBeLessThanOrEqual(3000); // 3s
   });
 
   // Justification: A store owner should be able to upload a discount, check that it is available,
   // and potentially delete it quickly.
   test("Single discount", async () => {
-    const start = Date.now();
+    // check add discount response time
+    var start = Date.now();
     const discount_response = await fetch(
       `${process.env.GATEWAY_BASE_URL}/discounts`,
       {
@@ -92,16 +114,22 @@ describe("Performance test", () => {
         },
         body: JSON.stringify({
           storeID: "test_store",
-          storeName: "Vancouver",
-          ingredient: "Toronto, the entire province",
+          storeName: "Test Store",
+          ingredient: "Apple",
           price: 1,
         }),
       }
     );
+    var end = Date.now();
+    var duration = end - start;
+    console.debug(`Add Discount Execution time: ${duration}ms`);
+    expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
 
     const data = await discount_response.json();
     const discountID = data.discountID;
 
+    // check get discount response time
+    start = Date.now();
     const get_response = await fetch(
       `${process.env.GATEWAY_BASE_URL}/discounts`,
       {
@@ -111,9 +139,15 @@ describe("Performance test", () => {
         },
       }
     );
+    end = Date.now();
+    duration = end - start;
+    console.debug(`Get Discount Execution time: ${duration}ms`);
+    expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
 
     expect(await get_response.json()).not.toBeNull();
 
+    // check delete discount response time
+    start = Date.now();
     const discount_teardown = await fetch(
       `${process.env.GATEWAY_BASE_URL}/discounts/${discountID}`,
       {
@@ -123,21 +157,24 @@ describe("Performance test", () => {
         },
       }
     );
+    end = Date.now();
     if (!discount_teardown.ok) {
       throw new Error(`HTTP error! Status: ${discount_teardown.status}`);
     }
-    const end = Date.now();
-    const duration = end - start; //begin timing test assuming that operation succeeded
-    console.debug(`Discount Execution time: ${duration}ms`);
-    expect(duration).toBeLessThanOrEqual(3000); // 3s
+    duration = end - start;
+    console.debug(`Delete Discount Execution time: ${duration}ms`);
+    expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
   });
 
   // Justification: A store owner should be able to upload multiple discounts, check that they are
   //  available, and delete any number of them quickly.
   test("10 discount", async () => {
-    const start = Date.now();
-    const discountID = [];
+    const addResponseTimes = [];
+    const discountIDs = [];
+
+    // check add discount response time on average over 10 iterations
     for (let i = 0; i < 10; i++) {
+      const start = Date.now();
       const discount_response = await fetch(
         `${process.env.GATEWAY_BASE_URL}/discounts`,
         {
@@ -147,16 +184,31 @@ describe("Performance test", () => {
           },
           body: JSON.stringify({
             storeID: "test_store",
-            storeName: "Vancouver",
-            ingredient: "Toronto, the entire province",
+            storeName: "Test Store",
+            ingredient: "Apple",
             price: i,
           }),
         }
       );
+      const end = Date.now();
+      const duration = end - start;
+      addResponseTimes.push(duration);
+
       const data = await discount_response.json();
       const discountIDsingle = data.discountID;
-      discountID.push(discountIDsingle);
+      discountIDs.push(discountIDsingle);
     }
+    const totalAddResponseTime = addResponseTimes.reduce(
+      (acc, time) => acc + time,
+      0
+    );
+    console.debug(
+      `Add 10 Discounts Average Execution time: ${totalAddResponseTime / 10}ms`
+    );
+    expect(totalAddResponseTime / 10).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
+
+    // check get all discounts response time
+    var start = Date.now();
     const get_response = await fetch(
       `${process.env.GATEWAY_BASE_URL}/discounts`,
       {
@@ -166,10 +218,24 @@ describe("Performance test", () => {
         },
       }
     );
+    var end = Date.now();
+
+    if (!get_response.ok) {
+      throw new Error(`HTTP error! Status: ${get_response.status}`);
+    }
+
+    var duration = end - start;
+    console.debug(`Get All Discounts Execution time: ${duration}ms`);
+    expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
+
+    // check delete discount response time on average over 10 iterations
+    const delResponseTimes = [];
     for (let i = 0; i < 10; i++) {
-      const did = discountID.pop();
+      const discountID = discountIDs.pop();
+
+      start = Date.now();
       const discount_teardown = await fetch(
-        `${process.env.GATEWAY_BASE_URL}/discounts/${did}`,
+        `${process.env.GATEWAY_BASE_URL}/discounts/${discountID}`,
         {
           method: "DELETE",
           headers: {
@@ -177,23 +243,32 @@ describe("Performance test", () => {
           },
         }
       );
+      end = Date.now();
+
       if (!discount_teardown.ok) {
         throw new Error(`HTTP error! Status: ${discount_teardown.status}`);
       }
+
+      duration = end - start;
+      delResponseTimes.push(duration);
     }
 
-    expect(await get_response.json()).toHaveLength(10); // assumes db was empty
-
-    const end = Date.now();
-    const duration = end - start; //begin timing test assuming that operation succeeded
-    console.debug(`10 discount Execution time: ${duration}ms`);
-    expect(duration).toBeLessThanOrEqual(3000); // 3s
+    const totalDelResponseTime = delResponseTimes.reduce(
+      (acc, time) => acc + time,
+      0
+    );
+    console.debug(
+      `Delete 10 Discounts Average Execution time: ${
+        totalDelResponseTime / 10
+      }ms`
+    );
+    expect(totalDelResponseTime / 10).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
   });
 
   // Justification: Users should be able to subscribe, and potentially change their mind quickly
   test("Single notification", async () => {
-    const userID = "real_person";
-    const start = Date.now();
+    // check add notification token response time
+    var start = Date.now();
     const notif_response = await fetch(
       `${process.env.GATEWAY_BASE_URL}/notifications`,
       {
@@ -202,16 +277,23 @@ describe("Performance test", () => {
           "Content-Type": "application/json",
         },
         body: JSON.stringify({
-          userID,
+          userID: USER_ID,
           fcmToken: "Vancouver",
         }),
       }
     );
+    var end = Date.now();
+    var duration = end - start;
+    console.debug(`Add Notification Execution time: ${duration}ms`);
+    expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
+
     const data = await notif_response.json();
     expect(data.message).toContain("Subscribed successfully");
 
+    // check delete notification token response time
+    start = Date.now();
     const notif_teardown = await fetch(
-      `${process.env.GATEWAY_BASE_URL}/notifications/${userID}`,
+      `${process.env.GATEWAY_BASE_URL}/notifications/${USER_ID}`,
       {
         method: "DELETE",
         headers: {
@@ -219,22 +301,26 @@ describe("Performance test", () => {
         },
       }
     );
+    end = Date.now();
+
     if (!notif_teardown.ok) {
       throw new Error(
         `notification delete error! Status: ${notif_teardown.status}`
       );
     }
-    const end = Date.now();
-    const duration = end - start; //begin timing test assuming that operation succeeded
-    console.debug(`notif Execution time: ${duration}ms`);
-    expect(duration).toBeLessThanOrEqual(3000); // 3s
+
+    duration = end - start;
+    console.debug(`Delete Notification Execution time: ${duration}ms`);
+    expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
   });
 
   // Justification: Multiple users may subscribe and potentially change their minds, which should
   // not cause significant delays
   test("10 notification", async () => {
-    const start = Date.now();
+    // check add notification token response time on average over 10 iterations
+    const addResponseTimes = [];
     for (let i = 0; i < 10; i++) {
+      const start = Date.now();
       const notif_response = await fetch(
         `${process.env.GATEWAY_BASE_URL}/notifications`,
         {
@@ -248,10 +334,28 @@ describe("Performance test", () => {
           }),
         }
       );
+      const end = Date.now();
+      const duration = end - start;
+      addResponseTimes.push(duration);
+
       const data = await notif_response.json();
       expect(data.message).toContain("Subscribed successfully");
     }
+    const totalAddResponseTime = addResponseTimes.reduce(
+      (acc, time) => acc + time,
+      0
+    );
+    console.debug(
+      `Add 10 Notifications Average Execution time: ${
+        totalAddResponseTime / 10
+      }ms`
+    );
+    expect(totalAddResponseTime / 10).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
+
+    // check delete notification token response time on average over 10 iterations
+    const delResponseTimes = [];
     for (let i = 0; i < 10; i++) {
+      const start = Date.now();
       const notif_teardown = await fetch(
         `${process.env.GATEWAY_BASE_URL}/notifications/${i}`,
         {
@@ -261,23 +365,35 @@ describe("Performance test", () => {
           },
         }
       );
+      const end = Date.now();
+
       if (!notif_teardown.ok) {
         throw new Error(
           `10 notification delete error! Status: ${notif_teardown.status}`
         );
       }
+
+      const duration = end - start;
+      delResponseTimes.push(duration);
     }
-    const end = Date.now();
-    const duration = end - start; //begin timing test assuming that operation succeeded
-    console.debug(`10 notification Execution time: ${duration}ms`);
-    expect(duration).toBeLessThanOrEqual(3000); // 3s
+    const totalDelResponseTime = delResponseTimes.reduce(
+      (acc, time) => acc + time,
+      0
+    );
+    console.debug(
+      `Delete 10 Notifications Average Execution time: ${
+        totalDelResponseTime / 10
+      }ms`
+    );
+    expect(totalDelResponseTime / 10).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
   });
 
   // Justification: Users should be able to add allergens, view them, and potentially delete them quickly
   test("Single allergy", async () => {
-    const userID = "real_person";
-    const allergen = "fake_people";
-    const start = Date.now();
+    const allergen = "fake_allergen";
+
+    // check add allergy response time
+    var start = Date.now();
     const allergy_response = await fetch(
       `${process.env.GATEWAY_BASE_URL}/preferences/allergies/`,
       {
@@ -286,15 +402,23 @@ describe("Performance test", () => {
           "Content-Type": "application/json",
         },
         body: JSON.stringify({
-          userID,
+          userID: USER_ID,
           allergy: allergen,
         }),
       }
     );
+    var end = Date.now();
+    var duration = end - start;
+    console.debug(`Add Allergy Execution time: ${duration}ms`);
+    expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
+
     const data = await allergy_response.json();
     expect(data.message).toContain("Allergy added successfully");
+
+    // check get allergy response time
+    start = Date.now();
     const get_response = await fetch(
-      `${process.env.GATEWAY_BASE_URL}/preferences/allergies/${userID}`,
+      `${process.env.GATEWAY_BASE_URL}/preferences/allergies/${USER_ID}`,
       {
         method: "GET",
         headers: {
@@ -302,10 +426,17 @@ describe("Performance test", () => {
         },
       }
     );
+    end = Date.now();
+    duration = end - start;
+    console.debug(`Get Allergy Execution time: ${duration}ms`);
+    expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
+
     expect(await get_response.json()).not.toBeNull();
-    const end = Date.now();
+
+    // check delete allergy response time
+    start = Date.now();
     const allergy_teardown = await fetch(
-      `${process.env.GATEWAY_BASE_URL}/preferences/allergies/${userID}/${allergen}`,
+      `${process.env.GATEWAY_BASE_URL}/preferences/allergies/${USER_ID}/${allergen}`,
       {
         method: "DELETE",
         headers: {
@@ -313,13 +444,16 @@ describe("Performance test", () => {
         },
       }
     );
+    end = Date.now();
+
     if (!allergy_teardown.ok) {
       throw new Error(
         `allergy delete error! Status: ${allergy_teardown.status}`
       );
     }
-    const duration = end - start; //begin timing test assuming that operation succeeded
-    console.debug(`allergy Execution time: ${duration}ms`);
-    expect(duration).toBeLessThanOrEqual(3000); // 3s
+
+    duration = end - start;
+    console.debug(`Delete Allergy Execution time: ${duration}ms`);
+    expect(duration).toBeLessThanOrEqual(MAX_RESPONSE_TIME);
   });
 });
